random: one generator, seeded by the machine, entropy from the OS
Three host_rand_64 implementations and six host_entropy_64 become one stream
in core/sys/random.c and one seam:

    core/sys/random.h  sys_random()      the machine's stream
                       sys_random_step() a stream of one's own
    host/host.h        host_random_seed() the machine answers, always valid
    osal/os.h          os_random_seed()   the OS's entropy, on request

sys_random asks the machine once on first use. A machine answers from a
fixture, from --seed, or by asking the OS -- so there is no sentinel and no
question of who decides. attr calls sys_random directly; the guest only ever
saw 31 bits of it.

Thirty-two bits, so a soft CPU stops running 64-bit multiplies to discard
half the result. Seed zero used to mean "unseeded" and was mapped to 1; the
increment is odd, so it is now an ordinary state, which matters because a
fixture may pin it.

The plan predicted the RIA would shrink. It grows 104 bytes instead: it never
had a generator -- it called the hardware RNG per draw -- so uniformity costs
it one. That was the trade, and the hardware is still spent, once, on the
seed. Everything that did have a generator shrinks: pocket -56, libretro
-4072, emu -56, wasm -135.

Four tests replace what the plan left as a hand check, because the screenshot
check I tried was a null result -- the ROM's output does not depend on the
fill, so it passed for the wrong seed too.

The bench owes host_random_seed like any machine, and every test that pulls
attr.c needs it, so tb_seed.c is linked into all of them rather than named at
four call sites.

+++ src/core/mem/mem.h
@@ -25,7 +25,7 @@ extern uint8_t ram[0x10000];
  * declare xram __uninitialized_ram() — so random is the default and a program
  * reading a byte it never wrote fails here instead of only on a Pico. Config,
  * set before mem_init adopts it; the seed is the run's, so the fill repeats. */
-void mem_set_fill(bool random, uint8_t value, uint64_t seed);
+void mem_set_fill(bool random, uint8_t value, uint32_t seed);
 
 /* Fill memory. First of the drivers, so what the ROM loader writes lands on top
  * of the fill, which is the order the hardware does it in. */
